Prune duplicates from each DNSBL file
The logging option is hard-coded disabled; enable it by editing
pfblockerng_extra.inc. Perhaps a GUI option to enable logging for this
feature is worthwhile.

Checkbox on the DNSBL page to enable this feature.

With lists like HaGeZi Pro, OISD big, AnudeepND Ads, StevenBlack,
malware-filter's urlhaus-filter and phishing list, AdAway, Disconnect
Ad, Disconnect Tracking, Disconnect Malvertising, YoYo, memory peaked at
just under 500MB according to top on an 8GB quad core machine.
~490MB for 'standard'
~430MB for 'pointer'

The Python script included can run standalone. Pass a directory
containing .txt files that have been run through pfBlockerNG's parsing
and it will prune the individual files as a collective set of all files
and write back to the same file names with a .pruned extension.

This can aid in evaluating the value of a list. That is, how many unique
domains among a set of lists does a list have? Is the list a near
overlap, complete overlap or mostly unique? Another possibility is to
peruse the pruned lists and identify domains that may be suitable for
regex blacklist.

Memory usage of Unbound is not significantly less after pruning.

Time to prune is about 40 seconds for a set as described above on an
Atom E3845 @ 1.91GHz with 8GB RAM.

TLD is enabled.
Prune is disabled.
Unbound using ~222MB of memory.

TLD is enabled
Prune is enabled.
Unbound using ~213MB of memory.

TLD is disabled
Prune is enabled.
Unbound using ~215MB of memory.

TLD is disabled
Prune is disabled.
Unbound using ~222MB of memory.

skip built-in prune. rely on C impl'ed solution